+++ modules/nebula4scala-cats-effect/src/main/scala/nebula4scala/cats/effect/NebulaSessionImpl.scala
@@ -1,4 +1,4 @@
-package nebula4scala.cats
+package nebula4scala.cats.effect
 
 import cats.effect.kernel.Async
 import cats.syntax.all._
@@ -6,7 +6,7 @@ import cats.syntax.all._
 import com.vesoft.nebula.client.graph.data.HostAddress
 
 import nebula4scala.api._
-import nebula4scala.cats.syntax._
+import nebula4scala.cats.effect.syntax._
 import nebula4scala.data.input._
 import nebula4scala.syntax._
